v1/requirements.txt

black
flake8
flask
Flask-Cors
flask_sqlalchemy
numpy 
pandas
pillow
pycodestyle
pyflakes
pytest
python-dotenv
pyyaml

